requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-groq